Created script "movementComputer"
This script is similar to "movement," but adds the ability to turn using
the mouse and strife with the left and right movement keys as well as
look up. I tried making it lock the mouse to the center... that is
currently really bugged and I could figure out how to do it currently.
The code I had for it is commented out, I believe.