Added environment assets from Unity, added sewer scene
It took me a lot longer than I care to admit to figure out how to add water to the project. Along with the water that I've added, there are also assets for making trees, along with assets for dirt and grass textures, among maybe others.

For the sewer scene, it's basically a bunch of rectangles at this moment with water on the lowest level. I didn't add a player character yet.