Improved water and lighting in World 2-2
Changed some properties of the water in World 2-2, and giving it a dark green base color. Also changed the coloring of the walls, ceiling, and floor in the world to make it seem more like a sewer. Also added lights and light fixture objects, and removed the "sun" in World 2-2.